round of cleanup
