+++ packages/lesswrong/testing/client.tests.js
@@ -1,4 +1,3 @@
 import '../components/comments/CommentsItem/_comments-unit-tests.js';
-import './moderation/moderation.client.tests.js';
-import '../components/comments/CommentsItem/_comments-unit-tests.js';
 import '../components/posts/timeframeUtils.tests.js';
+import './moderation/moderation.client.tests.js';
